Backed out 2 changesets (bug 1912695) for causing bustages on WasmBCMemory.cpp CLOSED TREE
Backed out changeset 7961243343b2 (bug 1912695)
Backed out changeset 4fa0b44a5701 (bug 1912695)

UltraBlame original commit: a5479981d97be54d32cad0c67a8228172d0a6d46

+++ js/src/wasm/WasmIonCompile.cpp
@@ -1361,7 +1361,7 @@ class FunctionCompiler {
   
   
   void foldConstantPointer(MemoryAccessDesc* access, MDefinition** base) {
-    uint64_t offsetGuardLimit = GetMaxOffsetGuardLimit(
+    uint32_t offsetGuardLimit = GetMaxOffsetGuardLimit(
         codeMeta_.hugeMemoryEnabled(access->memoryIndex()));
 
     if ((*base)->isConstant()) {
@@ -1373,8 +1373,9 @@ class FunctionCompiler {
       }
 
       uint64_t offset = access->offset64();
+
       if (offset < offsetGuardLimit && basePtr < offsetGuardLimit - offset) {
-        offset += basePtr;
+        offset += uint32_t(basePtr);
         access->setOffset32(uint32_t(offset));
         *base = isMem64(access->memoryIndex()) ? constantI64(int64_t(0))
                                                : constantI32(0);
@@ -1386,7 +1387,7 @@ class FunctionCompiler {
   
   void maybeComputeEffectiveAddress(MemoryAccessDesc* access,
                                     MDefinition** base, bool mustAddOffset) {
-    uint64_t offsetGuardLimit = GetMaxOffsetGuardLimit(
+    uint32_t offsetGuardLimit = GetMaxOffsetGuardLimit(
         codeMeta_.hugeMemoryEnabled(access->memoryIndex()));
 
     if (access->offset64() >= offsetGuardLimit ||
@@ -1540,7 +1541,6 @@ class FunctionCompiler {
     return codeMeta_.hugeMemoryEnabled(memoryIndex);
   }
 
-  
   
   MDefinition* computeEffectiveAddress(MDefinition* base,
                                        MemoryAccessDesc* access) {
